Rewrite admin fns
