--- functions/src/attribution.ts
@@ -1,35 +1,0 @@
-import { firestore } from "./config/firebase";
-import * as functions from "firebase-functions";
-import { AttributionEntry } from "./types";
-
-function compare(a: AttributionEntry, b: AttributionEntry) {
-  const aT = a.text.toLowerCase().replace(/[^a-z0-9]/, "");
-  const bT = b.text.toLowerCase().replace(/[^a-z0-9]/, "");
-  if (aT < bT) {
-    return -1;
-  }
-  if (aT > bT) {
-    return 1;
-  }
-  return 0;
-}
-
-export const addAttribution = functions.https.onCall(
-  async (entry: AttributionEntry, context: functions.https.CallableContext) => {
-    if (!context.auth) {
-      throw new functions.https.HttpsError(
-        "failed-precondition",
-        "Authentication required"
-      );
-    }
-
-    const docRef = await firestore.collection("attribution").doc("images");
-    const attributionEntries = (await docRef.get()).get("entries");
-    attributionEntries.push({
-      text: entry.text,
-      href: entry.href,
-    });
-    attributionEntries.sort(compare);
-    await docRef.update("entries", attributionEntries);
-  }
-);
