+++ functions/src/index.ts
@@ -1,6 +1,3 @@
-import { addAttribution } from "./attribution";
-import { uploadEntry } from "./admin";
-
 import { updateRssOnChange, serveRss, serveStagedRss } from "./rss";
 import { updateWeb1OnChange, serveWeb1 } from "./web1";
 
@@ -10,7 +7,3 @@ exports.serveStagedRss = serveStagedRss;
 
 exports.updateWeb1OnChange = updateWeb1OnChange;
 exports.serveWeb1 = serveWeb1;
-
-// Admin
-exports.uploadEntry = uploadEntry;
-exports.addAttribution = addAttribution;
